pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "carriercapture"
version = "1.0.0"
description = "Carrier capture and non-radiative recombination rate calculations for semiconductors"
authors = [
    {name = "CarrierCapture Team"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21",
    "scipy>=1.7",
    "pandas>=1.3",
    "pyyaml>=5.4",
    "click>=8.0",
    "rich>=10.0",
    "plotly>=5.0",
    "dash>=2.0",
    "pymatgen>=2022.0",
    "joblib>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "pytest-xdist>=2.5",
    "mypy>=0.990",
    "black>=22.0",
    "ruff>=0.0.250",
    "pre-commit>=2.20",
]
docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
    "nbsphinx>=0.8",
    "myst-parser>=0.18",
]
notebook = [
    "jupyter>=1.0",
    "ipywidgets>=8.0",
]
doped = [
    "doped>=2.0",
    "monty>=2024.0",
]

[project.scripts]
carriercapture = "carriercapture.cli.main:cli"

[project.urls]
Homepage = "https://github.com/yourorg/CarrierCapture.py"
Repository = "https://github.com/yourorg/CarrierCapture.py"

[tool.hatch.build.targets.wheel]
packages = ["src/carriercapture"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=carriercapture --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
